requirements.txt
psycopg2-binary>=2.9
streamlit>=1.16
sqlalchemy>=1.4.0